Small css changes, name change to Filmoteka

+++ app/controllers/pages_controller.rb
@@ -4,11 +4,11 @@ class PagesController < ApplicationController
 	end
 
 	def movies
-		@movies = Movie.where(show: false).sort {|a,b| b.overal_score.to_i <=> a.overal_score.to_i}[0..9]
+		@movies = Movie.where(show: false).sort {|a,b| b.overal_score.to_f <=> a.overal_score.to_f}[0..9]
 	end
 
 	def shows
-		@shows = Movie.where(show: true).sort {|a,b| b.overal_score.to_i <=> a.overal_score.to_i}[0..9]
+		@shows = Movie.where(show: true).sort {|a,b| b.overal_score.to_f <=> a.overal_score.to_f}[0..9]
 	end
 
 end
